+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [3.2.1] - 2025-11-03
+
 ### Fixed
 
 - Handle null offering rules on CourseProduct item
@@ -2367,7 +2369,8 @@ us:
 - finish integrating the missing pages and improve the sandbox environment;
 - test and polish the use of richie as a django app / node dependency.
 
-[unreleased]: https://github.com/openfun/richie/compare/v3.2.0...master
+[unreleased]: https://github.com/openfun/richie/compare/v3.2.1...master
+[3.2.0]: https://github.com/openfun/richie/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/openfun/richie/compare/v3.1.2...v3.2.0
 [3.1.2]: https://github.com/openfun/richie/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/openfun/richie/compare/v3.1.0...v3.1.1